pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "google-genai>=1.63.0",
    "nltk==3.9.1",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
]
